Start Dal Create Type <=> Add new table Add some intents Need to fix Dialog choice at data selector TYPE
PROBLEM STOPWATCH RESETS 10 SECONDS!!!
Database works but needs a look up on a diffrent problem of it

add 2 new Dals

add users class

started on 2 new dals

+++ app/src/main/java/com/example/fixedproject/Type.java
@@ -1,17 +1,17 @@
 package com.example.fixedproject;
 
 public class Type {
-    private int tid;
+    private long tid;
     private String tname;
-    public Type(String tname, int tid){ this.tname = tname; this.tid = tid; }
+    public Type(String tname, long tid){ this.tname = tname; this.tid = tid; }
     public String getTname(){
         return tname;
     }
     public void setTname(String tname){ this.tname = tname; }
-    public int getTid(){
+    public long getTid(){
         return tid;
     }
-    public void setTid(int tid){
+    public void setTid(long tid){
         this.tid=tid;
     }
 }
